New feature: run modes
With this, hiera_include will only query for a special set of classes
to include if the runmode fact is set. This enables e.g bootstrap runs
with the same hiera data set as used for the 'full' runs:

FACTER_RUNMODE=bootstrap puppet apply site.pp

This will include all classes specified as classes_bootstrap in Hiera.