Update bgo-compute-vgpu-53.yaml - blacklist nouveau
Problem when using Nvidia driver R535 (16) for A100 cards, it is not supported but AIE driver R550 (17) works.  Last know LTS version for A100 is 13 or R470. 

I just want to try blacklist nouveau to see if it behaves differently. 

NVRM: Aborting probe for VF 0000:3b:02.3 since PF is not bound to nvidia driver.
nvidia: probe of 0000:3b:02.3 failed with error -1